requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
Django==4.1.7
django-cors-headers==3.14.0
django-crum==0.7.9
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-gis==1.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.0
Fiona==1.9.1
geopandas==0.12.2
inflection==0.5.1
jsonschema==4.17.3
munch==2.5.0
numpy==1.24.2
packaging==23.0
pandas==1.5.3
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyproj==3.4.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
shapely==2.0.1
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
